api/requirements.txt
fastapi[standard]
numpy
Pillow
